feat: :recycle: Split up CM views for more readability and moved all add, modify & delete to one url
