+++ Project/UWEFlix/urls.py
@@ -14,12 +14,8 @@ urlpatterns = [
     path('modify_account', views.modify_account),
     path('cinema_management', views.cm_dash, name='cinema_management'),
     path('film_management', views.film_dash, name='film_management'),
-    path('add_film', views.add_film, name='add_film'),
-    path('modify_film', views.modify_film, name='modify_film'),
-    path('delete_film', views.delete_film, name='delete_film'),
     path('showings_management', views.showings_dash, name='showings_management'),
-    path('add_showing', views.add_showing, name='add_showing'),
-    path('modify_showing', views.modify_showing, name='modify_showing'),
-    path('delete_showing', views.delete_showing, name='delete_showing'),
-
+    path('add', views.add, name='add'),
+    path('modify', views.modify, name='modify'),
+    path('delete', views.delete, name='delete'),
 ]
