+++ Project/UWEFlix/views/__init__.py
@@ -2,5 +2,4 @@ from .login import *
 from .booking import *
 from .account import *
 from .account_manager import *
-from .cinema_manager import *
-from .showings_management import *
+from .cinema_management import *
